requirements.txt
asyncpg==0.22.0
async-timeout==3.0.1
Cybernator==0.6.3
discord.py==1.7.1
postgres==3.0.0
pip==21.1.1
typing-extensions==3.7.4.3
environs==9.3.0
asyncio==3.4.3